feat(task): update source insertion to mimick AngularJS's ng-include original behavior
When AngularJS finds an ng-include directive, it fetches the source and puts it inside the element (as its innerHTML).
This commit makes the nginclude task follow this procedure. Previously, it was removing the element that had the ng-include directive and placing the contents in its place. This could cause issues with missing elements, such as styling problems.